🎨 lbaas/v1: unify state attribute and add getters
Add some easy to use getters for the state of LBaaS resources.
