+++ pkg/apis/lbaas/v1/backend_types.go
@@ -4,6 +4,8 @@ package v1
 
 // The Backend resource configures settings common for all specific backend Server resources linked to it.
 type Backend struct {
+	HasState
+
 	CustomerIdentifier string     `json:"customer_identifier"`
 	ResellerIdentifier string     `json:"reseller_identifier"`
 	Identifier         string     `json:"identifier" anxcloud:"identifier"`
@@ -11,7 +13,6 @@ type Backend struct {
 	HealthCheck        string     `json:"health_check"`
 	Mode               Mode       `json:"mode"`
 	ServerTimeout      int        `json:"server_timeout"`
-	State              State      `json:"state"`
 	AutomationRules    []RuleInfo `json:"automation_rules,omitempty"`
 
 	// Only the name and identifier fields are used and returned.
